protect game api endpoints
