+++ backend/srcs/backend/conf/games/urls.py
@@ -1,11 +1,11 @@
 from django.urls import path
-from .views import (
-    ListGamesView,
-    # GameDetailsView,
-    RetrieveUpdateGameView,
-)
+# from .views import (
+#     # ListGamesView,
+#     # GameDetailsView,
+#     # RetrieveUpdateGameView,
+# )
 
 urlpatterns = [
-    path("<uuid:pk>/games/edit/", RetrieveUpdateGameView.as_view(), name="game_update"),
+    # path("<uuid:pk>/games/edit/", RetrieveUpdateGameView.as_view(), name="game_update"),
     # path("<uuid:pk>/<id:pk>", ListGamesView.as_view(), name="game_list"),
 ]
